requirements.txt
absl-py==1.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blis==0.7.6
cached-property==1.5.2
cachetools==5.0.0
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cymem==2.0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docx==0.2.4
docx2txt==0.8
dominate==2.6.0
email-validator==1.1.3
entrypoints==0.4
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.45.0
gunicorn==20.1.0
h5py==3.6.0
heroku==0.1.4
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.10.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.0
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.10
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20220319
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
preshed==3.0.6
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
pyparsing==3.0.7
pyresparser==1.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==223
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spacy==2.3.5
SQLAlchemy==1.4.32
srsly==1.0.5
tensorflow-cpu==2.7.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thinc==7.4.5
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
visitor==0.1.3
waitress==2.1.1
wasabi==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.14.0
zipp==3.7.0
